+++ types/protocol-mapping.d.ts
@@ -309,6 +309,11 @@ export namespace ProtocolMapping {
          */
         'Network.webSocketWillSendHandshakeRequest': [Protocol.Network.WebSocketWillSendHandshakeRequestEvent];
         /**
+         * Fired upon WebTransport creation.
+         */
+        'Network.webTransportCreated': [Protocol.Network.WebTransportCreatedEvent];
+        'Network.webTransportClosed': [Protocol.Network.WebTransportClosedEvent];
+        /**
          * Fired when additional information about a requestWillBeSent event is available from the
          * network stack. Not every requestWillBeSent event will have an additional
          * requestWillBeSentExtraInfo fired for it, and there is no guarantee whether requestWillBeSent